+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* [refactor] Use Tutor v1 plugin API
+
 ## Version 0.0.6 (2022-04-19)
 
 * [feature] List available backup versions using `tutor k8s restore